requirements.txt
django_address==0.2.0
django_registration==3.0
Django==2.1.1
django_phonenumber_field==2.0.1
django_newsletter==0.7
address==0.1.1
django_crispy_forms==1.7.2
newsletter==0.1.22pre
djangorestframework==3.8.2
allauth==0.17.1
